+++ adapters/zemismart/ZML03EZ.py
@@ -1,0 +1,20 @@
+import Domoticz
+import json
+from adapters.base_adapter import Adapter
+from devices.switch.on_off_switch import OnOffSwitch
+
+class ZML03EZ(Adapter):
+    def __init__(self, devices):
+        super().__init__(devices)
+
+        self.devices.append(OnOffSwitch(devices, 'left', 'state_left'))
+        self.devices.append(OnOffSwitch(devices, 'center', 'state_center'))
+        self.devices.append(OnOffSwitch(devices, 'right', 'state_right'))
+
+    def handleCommand(self, alias, device, device_data, command, level, color):
+        return {
+            'topic': '/'.join([device_data['friendly_name'], alias, 'set']),
+            'payload': json.dumps({
+                "state": command.upper()
+            })
+        }